LayoutTests/imported/w3c/web-platform-tests/trusted-types/eval-csp-tt-no-default-policy-expected.txt

CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"

PASS eval of TrustedScript works.
PASS indirect eval of TrustedScript works.
PASS eval of string fails.
PASS indirect eval of string fails.
PASS eval of !TrustedScript and !string works.
PASS Function constructor of TrustedScript works.
FAIL Function constructor of string fails. assert_throws_js: function "_ => new Function('return 1+1')()" did not throw
PASS Function constructor of all TrustedScripts works.
FAIL Function constructor of all strings fails. assert_throws_js: function "_ => new Function('val', 'return val+1')(1)" did not throw
FAIL Function constructor of string and TrustedScript fails. assert_throws_js: function "_ => new Function('val', p.createScript('return val+1'))(1)" did not throw
FAIL AsyncFunction constructor of string fails. assert_throws_js: function "_ => new AsyncFunction('return 1+1')()" did not throw
FAIL GeneratorFunction constructor of string fails. assert_throws_js: function "_ => new GeneratorFunction('return 1+1')()" did not throw
FAIL AsyncGeneratorFunction constructor of string fails. assert_throws_js: function "_ => new AsyncGeneratorFunction('return 1+1')()" did not throw
